Fixed error with mulitple additions

+++ frontend/src/app/user-list/user-list.component.ts
@@ -38,7 +38,10 @@ export class UserListComponent implements OnInit {
   }
 
   deleteUser(user:User) : void{
-    this.userService.delete(+user.id).subscribe(d=>this.userService.gotoUserList()) 
+    let check = confirm("Are you sure you want to delete this employee?");
+    if (check) {
+      this.userService.delete(+user.id).subscribe(d=>this.userService.gotoUserList())
+    }
   }
 
   ngOnInit() {
